batch/requirements.txt
backports-abc==0.4
decorator==4.0.10
entrypoints==0.2.2
GDAL==2.0.1
geographiclib==1.46.3
h5py==2.6.0
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
Keras==1.0.8
MarkupSafe==0.23
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.3
numpy==1.11.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
PyYAML==3.12
pyzmq==15.4.0
qtconsole==4.2.1
readline==6.2.4.1
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
Theano==0.8.2
tornado==4.4.1
traitlets==4.3.0
wcwidth==0.1.7
widgetsnbextension==1.2.6
